+++ module/ASC.Data.Reassigns/RemoveProgressItem.cs
@@ -35,7 +35,7 @@ using ASC.Common.Threading.Progress;
 using ASC.Core;
 using ASC.Core.Users;
 using ASC.Data.Storage;
-using ASC.Mail.GarbageEraser;
+using ASC.Mail.Core.Engine;
 using ASC.MessagingSystem;
 using ASC.Web.CRM.Core;
 using ASC.Web.Core;
@@ -55,9 +55,10 @@ namespace ASC.Data.Reassigns
         private readonly Guid _userId;
         private readonly string _userName;
         private readonly Guid _currentUserId;
+        private readonly bool _notify;
 
         private readonly IFileStorageService _docService;
-        private readonly MailGarbageEraser _mailEraser;
+        private readonly MailGarbageEngine _mailEraser;
 
         public object Id { get; set; }
         public object Status { get; set; }
@@ -66,7 +67,7 @@ namespace ASC.Data.Reassigns
         public bool IsCompleted { get; set; }
         public Guid FromUser { get { return _userId; } }
 
-        public RemoveProgressItem(HttpContext context, int tenantId, UserInfo user, Guid currentUserId)
+        public RemoveProgressItem(HttpContext context, int tenantId, UserInfo user, Guid currentUserId, bool notify)
         {
             _context = context;
             _httpHeaders = QueueWorker.GetHttpHeaders(context.Request);
@@ -75,9 +76,10 @@ namespace ASC.Data.Reassigns
             _userId = user.ID;
             _userName = UserFormatter.GetUserName(user, DisplayUserNameFormat.Default);
             _currentUserId = currentUserId;
+            _notify = notify;
 
             _docService = Web.Files.Classes.Global.FileStorageService;
-            _mailEraser = new MailGarbageEraser();
+            _mailEraser = new MailGarbageEngine();
 
             Id = QueueWorker.GetProgressItemId(tenantId, _userId, typeof(RemoveProgressItem));
             Status = ProgressStatus.Queued;
@@ -108,14 +110,21 @@ namespace ASC.Data.Reassigns
                 Percentage = 25;
                 _docService.DeleteStorage(_userId);
 
-                logger.Info("deleting of data from crm");
-
-                Percentage = 50;
-                using (var scope = DIHelper.Resolve(_tenantId))
+                if (!CoreContext.Configuration.CustomMode)
                 {
-                    var crmDaoFactory = scope.Resolve<CrmDaoFactory>();
-                    crmSpace = crmDaoFactory.ReportDao.GetFiles(_userId).Sum(file => file.ContentLength);
-                    crmDaoFactory.ReportDao.DeleteFiles(_userId);
+                    logger.Info("deleting of data from crm");
+
+                    Percentage = 50;
+                    using (var scope = DIHelper.Resolve(_tenantId))
+                    {
+                        var crmDaoFactory = scope.Resolve<CrmDaoFactory>();
+                        crmSpace = crmDaoFactory.ReportDao.GetFiles(_userId).Sum(file => file.ContentLength);
+                        crmDaoFactory.ReportDao.DeleteFiles(_userId);
+                    }
+                }
+                else
+                {
+                    crmSpace = 0;
                 }
 
                 logger.Info("deleting of data from mail");
@@ -208,22 +217,24 @@ namespace ASC.Data.Reassigns
 
         private void SendSuccessNotify(long docsSpace, long crmSpace, long mailSpace, long talkSpace)
         {
-            StudioNotifyService.Instance.SendMsgRemoveUserDataCompleted(_currentUserId, _userId, _userName, docsSpace, crmSpace, mailSpace, talkSpace);
+            if (_notify)
+                StudioNotifyService.Instance.SendMsgRemoveUserDataCompleted(_currentUserId, _userId, _userName,
+                                                                            docsSpace, crmSpace, mailSpace, talkSpace);
 
             if (_httpHeaders != null)
-            {
-                MessageService.Send(_httpHeaders, MessageAction.UserDataRemoving, new[] { _userName });
-            }
+                MessageService.Send(_httpHeaders, MessageAction.UserDataRemoving, MessageTarget.Create(_userId),
+                                    new[] {_userName});
             else
-            {
-                MessageService.Send(_context.Request, MessageAction.UserDataRemoving, _userName);
-            }
+                MessageService.Send(_context.Request, MessageAction.UserDataRemoving, MessageTarget.Create(_userId),
+                                    _userName);
         }
 
         private void SendErrorNotify(string errorMessage)
         {
-            StudioNotifyService.Instance.SendMsgRemoveUserDataFailed(_currentUserId, _userId, _userName, errorMessage);
-        }
+            if (!_notify) return;
 
+            StudioNotifyService.Instance.SendMsgRemoveUserDataFailed(_currentUserId, _userId, _userName,
+                                                                     errorMessage);
+        }
     }
 }
