+++ module/ASC.Files.Thirdparty/ProviderDao/ProviderDaoBase.cs
@@ -27,6 +27,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using ASC.Files.Core;
 using ASC.Files.Core.Data;
 using ASC.Files.Core.Security;
@@ -36,7 +37,9 @@ using ASC.Files.Thirdparty.GoogleDrive;
 using ASC.Files.Thirdparty.OneDrive;
 using ASC.Files.Thirdparty.SharePoint;
 using ASC.Files.Thirdparty.Sharpbox;
+using ASC.Web.Files.Resources;
 using ASC.Web.Files.Utils;
+using ASC.Web.Studio.Core;
 
 namespace ASC.Files.Thirdparty.ProviderDao
 {
@@ -176,6 +179,12 @@ namespace ASC.Files.Thirdparty.ProviderDao
             var toFileDao = toSelector.GetFileDao(toFolderId);
             var fromFile = fromFileDao.GetFile(fromSelector.ConvertId(fromFileId));
 
+            if (fromFile.ContentLength > SetupInfo.AvailableFileSize)
+            {
+                throw new Exception(string.Format(deleteSourceFile ? FilesCommonResource.ErrorMassage_FileSizeMove : FilesCommonResource.ErrorMassage_FileSizeCopy,
+                                                  FileSizeComment.FilesSizeToString(SetupInfo.AvailableFileSize)));
+            }
+
             using (var securityDao = TryGetSecurityDao())
             using (var tagDao = TryGetTagDao())
             {
@@ -184,7 +193,7 @@ namespace ASC.Files.Thirdparty.ProviderDao
 
                 var toFile = toFileDao.GetFile(toSelector.ConvertId(toFolderId), fromFile.Title);
 
-                if (toFile == null)
+                if (toFile == null || deleteSourceFile)
                 {
                     fromFile.ID = fromSelector.ConvertId(fromFile.ID);
 
@@ -225,7 +234,7 @@ namespace ASC.Files.Thirdparty.ProviderDao
             }
         }
 
-        protected Folder PerformCrossDaoFolderCopy(object fromFolderId, object toRootFolderId, bool deleteSourceFolder)
+        protected Folder PerformCrossDaoFolderCopy(object fromFolderId, object toRootFolderId, bool deleteSourceFolder, CancellationToken? cancellationToken)
         {
             //Things get more complicated
             var fromSelector = GetSelector(fromFolderId);
@@ -249,15 +258,32 @@ namespace ASC.Files.Thirdparty.ProviderDao
                                          });
 
             var foldersToCopy = fromFolderDao.GetFolders(fromSelector.ConvertId(fromFolderId));
-            var filesToCopy = fromFileDao.GetFiles(fromSelector.ConvertId(fromFolderId));
+            var fileIdsToCopy = fromFileDao.GetFiles(fromSelector.ConvertId(fromFolderId));
+            Exception copyException = null;
             //Copy files first
-            foreach (var file in filesToCopy)
+            foreach (var fileId in fileIdsToCopy)
             {
-                PerformCrossDaoFileCopy(file, toFolderId, deleteSourceFolder);
+                if (cancellationToken.HasValue) cancellationToken.Value.ThrowIfCancellationRequested();
+                try
+                {
+                    PerformCrossDaoFileCopy(fileId, toFolderId, deleteSourceFolder);
+                }
+                catch (Exception ex)
+                {
+                    copyException = ex;
+                }
             }
             foreach (var folder in foldersToCopy)
             {
-                PerformCrossDaoFolderCopy(folder.ID, toFolderId, deleteSourceFolder);
+                if (cancellationToken.HasValue) cancellationToken.Value.ThrowIfCancellationRequested();
+                try
+                {
+                    PerformCrossDaoFolderCopy(folder.ID, toFolderId, deleteSourceFolder, cancellationToken);
+                }
+                catch (Exception ex)
+                {
+                    copyException = ex;
+                }
             }
 
             if (deleteSourceFolder)
@@ -289,11 +315,13 @@ namespace ASC.Files.Thirdparty.ProviderDao
                     }
                 }
 
-                fromFolderDao.DeleteFolder(fromSelector.ConvertId(fromFolderId));
-
+                if (copyException == null)
+                    fromFolderDao.DeleteFolder(fromSelector.ConvertId(fromFolderId));
             }
 
-            return toFolderDao.GetFolder(toFolderId);
+            if (copyException != null) throw copyException;
+
+            return toFolderDao.GetFolder(toSelector.ConvertId(toFolderId));
         }
     }
 }
